4 - jira agent/pyproject.toml
[project]
name = "4-jira-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
